@@ -1,8 +1,9 @@
 # coding: us-ascii
+# frozen_string_literal: true
+
 # Copyright (c) 2012 Kenichi Kamiya
 
 require_relative 'declare/version'
-require_relative 'declare/errors'
 require_relative 'declare/singleton_class'
 require_relative 'declare/assertions'
 require_relative 'declare/callerentry'
