@@ -1,4 +1,5 @@
 # coding: us-ascii
+# frozen_string_literal: true
 
 module Declare
   module ObjectExtension
@@ -6,16 +7,14 @@ module Declare
 
     # @return [Scope]
     def The(target, &block)
-      Declare.scope! target, _declare_called_from
+      Declare.scope!(target, _declare_called_from)
       Declare.new_scope(target, &block)
     end
-
-    alias_method :_original_caller, :caller
 
     # @param [Integer] level
     # @return [CallerEntry]
     def _declare_called_from(level=0)
-      CallerEntry.parse _original_caller[1 + level]
+      CallerEntry.parse(caller[1 + level])
     end
   end
 end
