@@ -1,8 +1,12 @@
 # coding: us-ascii
-require 'test/unit'
+# frozen_string_literal: true
+
+require_relative 'helper'
 
 class Test_Decalre_Behavior_Exception_Outer_NestedDeclare < Test::Unit::TestCase
-  CODE = <<"EOD"
+  include DeclareHelpers
+
+  CODE = <<"RUBY"
 class Person
   attr_reader :name, :birth
 
@@ -13,7 +17,7 @@ class Person
 
 end
 
-app_dir = "#{File.expand_path('..', File.dirname(__FILE__))}"
+app_dir = "#{File.expand_path('..', __dir__)}"
 require app_dir + '/lib/declare/autorun'
 
 The Person.new('John') do |john|
@@ -25,14 +29,11 @@ The Person.new('John') do |john|
 
   raise Exception
 end
-EOD
+RUBY
 
   def test_result
-    out = `ruby -w 2>&1 <<EOD
-#{CODE}
-EOD
-`
-    assert_equal(1, $?.exitstatus)
-    assert_match(/Exception/, out)
+    result = execute_ruby_code_in_other_process(CODE)
+    assert_match(/Exception/, result.all_output)
+    assert_equal(1, result.exitstatus)
   end
 end
